requirements.txt
fastapi==0.119.0
pydantic==2.12.2
python-dotenv==1.1.1
passlib==1.7.4
uvicorn==0.37.0
bcrypt==5.0.0
google-cloud-storage==3.4.1
google-genai==1.44.0
sqlalchemy==2.0.44
pymysql==1.1.2
python-multipart==0.0.20
pycryptodome==3.23.0
watchdog==6.0.0
validate-email==1.3
email-validator==2.3.0
pymupdf==1.26.5
pypdf==3.17.4
numpy==2.2.6
chromadb==1.2.1
tabula-py==2.10.0
jpype1==1.6.0
pdfplumber==0.11.7
camelot-py[cv]==1.0.9
opencv-python==4.12.0.88
pandas==2.3.3
pytesseract==0.3.13